+++ Bravo/Source/Private/BravoMesh.cpp
@@ -176,7 +176,7 @@ void BravoMesh::Render()
 	{
 		// draw mesh
 		glBindVertexArray(VAO);
-			glDrawElements(GL_TRIANGLES, Indices.size(), GL_UNSIGNED_INT, 0);
+			glDrawElements(GL_TRIANGLES, (int32)Indices.size(), GL_UNSIGNED_INT, 0);
 		glBindVertexArray(0);
 		glActiveTexture(0);
 	}
